amend me again
Tagging samples during a stroke is faster, but leads to some glitching - which is exactly what we're trying to avoid.

Not tagging samples during a stroke & only storing the cleanup results in a tempsample instead of the "real" sample is slower, but gives lovely correct-looking results.

I also now reset the node and sample tags whenever we finish a stroke. This is also important for correctness - samples away from the surface can become invalid when we start a new stroke, so shouldn't remain tagged as known good.